refactor: merge config packages
