+++ internal/agent/cloud/common.go
@@ -6,11 +6,11 @@ import (
 	"time"
 
 	"github.com/AikidoSec/firewall-go/internal/agent/aikido_types"
+	"github.com/AikidoSec/firewall-go/internal/agent/config"
 	"github.com/AikidoSec/firewall-go/internal/agent/globals"
 	"github.com/AikidoSec/firewall-go/internal/agent/log"
 	"github.com/AikidoSec/firewall-go/internal/agent/ratelimiting"
 	"github.com/AikidoSec/firewall-go/internal/agent/utils"
-	"github.com/AikidoSec/firewall-go/internal/config"
 )
 
 func GetAgentInfo() aikido_types.AgentInfo {
