+++ internal/agent/config/globals.go
@@ -1,0 +1,5 @@
+package config
+
+const (
+	Version = "1.0.0"
+)
