--- internal/config/globals.go
@@ -1,12 +1,0 @@
-package config
-
-import (
-	"sync"
-)
-
-var ServiceConfig ServiceConfigData
-var ServiceConfigMutex sync.RWMutex
-
-const (
-	Version = "1.0.0"
-)
