+++ internal/http/on_init_request.go
@@ -3,7 +3,7 @@ package http
 import (
 	"context"
 
-	"github.com/AikidoSec/firewall-go/internal/config"
+	"github.com/AikidoSec/firewall-go/internal/agent/config"
 	"github.com/AikidoSec/firewall-go/internal/helpers"
 	"github.com/AikidoSec/firewall-go/internal/request"
 )
